FLV/E-RTMP経路の重複コードを共通化する
コードレビューで確定した整理項目5件に対応する。挙動の変更は無い。

EnhancedRTMPTagHeader:
- ExVideoTagHeader/ExAudioTagHeaderに1つずつあったSkipModEx/ReadFourCcを
  ExTagHeaderReaderへ集約する。ModExとFourCCの形式はE-RTMP仕様上
  コーデック非依存で音声・映像とも完全に同一だった。

BitReader:
- FLVToMKVとFLVToMPEG2TSに同一実装があったものを共有ファイルへ移す。

FLVContentFilterSinkBase:
- FLVToMKVContentFilterSinkとFLVToTSContentFilterSinkで90行以上重複していた
  キュー処理を抽象基底へ切り出す。派生側が与えるのは変換ループと
  チャンネル情報の3文字列のみ。フォルト時に下流をOnStopで止める手当てや
  IsProcessorAliveによる短絡も一箇所になる。

FLVParseBuffer.Trim:
- TryGetBufferが失敗した場合の分岐を削除する。bufferはnew MemoryStream()で
  生成しており内部配列は常に公開されているため到達しない。GetBuffer()の
  直接呼び出しに置き換えた。

テスト:
- FLVToMKVTests/FLVToMPEG2TSTests/FLVContentBufferTestsで重複していた
  タグ組み立てとCaptureSinkをFLVTestHelpersへ切り出す。F#はコンパイル順序が
  意味を持つのでfsprojでは3つのテストより前に置く。
- CaptureSinkはヘッダを連結(Header)とパケット単位(Headers/HeaderCount)の
  両方で取れる形に統合した。
