+++ PeerCastStation/PeerCastStation.Test/FLVToMKVTests.fs
@@ -9,23 +9,9 @@ open PeerCastStation.Core
 open PeerCastStation.FLV
 open PeerCastStation.FLV.AMF
 open PeerCastStation.FLV.RTMP
+open FLVTestHelpers
 
 // ---- helpers ----
-
-let private indexOf (haystack:byte[]) (needle:byte[]) =
-    if needle.Length=0 || haystack.Length<needle.Length then -1
-    else
-        let last = haystack.Length - needle.Length
-        let rec search i =
-            if i>last then -1
-            else
-                let mutable j = 0
-                while j<needle.Length && haystack.[i+j]=needle.[j] do j <- j+1
-                if j=needle.Length then i else search (i+1)
-        search 0
-
-let private contains (haystack:byte[]) (needle:byte[]) =
-    indexOf haystack needle >= 0
 
 let private countOf (haystack:byte[]) (needle:byte[]) =
     if needle.Length=0 then 0
@@ -40,8 +26,6 @@ let private countOf (haystack:byte[]) (needle:byte[]) =
 let private startsWith (haystack:byte[]) (needle:byte[]) =
     haystack.Length>=needle.Length &&
     Array.forall2 (=) (Array.sub haystack 0 needle.Length) needle
-
-let private ascii (s:string) = System.Text.Encoding.ASCII.GetBytes(s)
 
 // ---- EBMLWriter primitive unit tests ----
 
@@ -77,22 +61,6 @@ let ``WriteElement は ID とサイズVINT と payload を連結する`` () =
 
 // ---- 構造検証(ラウンドトリップ) ----
 
-/// FLVタグ(11バイトヘッダ + body + 4バイトPreviousTagSize)を組み立てる。
-let private makeTag (typ:int) (timestamp:int) (body:byte[]) =
-    let ds = body.Length
-    let header =
-        [| byte typ
-           byte (ds>>>16); byte (ds>>>8); byte ds
-           byte (timestamp>>>16); byte (timestamp>>>8); byte timestamp; byte (timestamp>>>24)
-           0uy; 0uy; 0uy |]
-    let tagsize = ds + 11
-    let footer = [| byte (tagsize>>>24); byte (tagsize>>>16); byte (tagsize>>>8); byte tagsize |]
-    Array.concat [ header; body; footer ]
-
-let private flvHeader =
-    // "FLV" v1 flags(audio+video) DataOffset=9 PreviousTagSize0=0
-    [| 0x46uy;0x4Cuy;0x56uy; 1uy; 0x05uy; 0uy;0uy;0uy;9uy; 0uy;0uy;0uy;0uy |]
-
 let private onMetaDataBodyOf (width:AMFValue) (height:AMFValue) =
     let dict = System.Collections.Generic.Dictionary<string, AMFValue>()
     dict.["width"]  <- width
@@ -102,26 +70,6 @@ let private onMetaDataBodyOf (width:AMFValue) (height:AMFValue) =
 
 let private onMetaDataBody (width:float) (height:float) =
     onMetaDataBodyOf (AMFValue(width)) (AMFValue(height))
-
-// 最小の avcC(中身は検証では問わない。先頭5バイトを除いた部分が CodecPrivate になる)
-let private avcC =
-    [| 1uy;0x42uy;0x00uy;0x1Fuy;0xFFuy;0xE1uy;0x00uy;0x04uy;0x67uy;0x42uy;0x00uy;0x1Fuy;0x01uy;0x00uy;0x04uy;0x68uy;0xCEuy;0x3Cuy;0x80uy |]
-
-type private CaptureSink() =
-    let header = System.Collections.Generic.List<byte>()
-    let content = System.Collections.Generic.List<byte>()
-    member val ChannelType : string = null with get, set
-    member _.Header = header.ToArray()
-    member _.Content = content.ToArray()
-    interface IContentSink with
-        member this.OnChannelInfo(ci) = this.ChannelType <- ci.ContentType
-        member _.OnChannelTrack(_) = ()
-        member _.OnContentHeader(c) = header.AddRange(c.Data.ToArray())
-        member _.OnContent(c) = content.AddRange(c.Data.ToArray())
-        member _.OnStop(_) = ()
-
-let private newContent (data:byte[]) =
-    Content(0, TimeSpan.Zero, 0L, data, 0, data.Length, PCPChanPacketContinuation.None)
 
 [<Fact>]
 let ``FLV(H264+AAC) を MKV に変換し EBML 構造が成立する`` () =
@@ -212,22 +160,7 @@ let ``onMetaData が無い場合は映像を除外し音声のみで構成する
 // ---- E-RTMP(enhanced タグ)経路 ----
 
 // ダミーのコーデック設定(構造検証では中身は問わない。CodecPrivate へ無加工で入ることだけ確認する)
-let private hvcC = [| 1uy;0x01uy;0x60uy;0x00uy;0x00uy;0x03uy;0x00uy;0x90uy;0x12uy;0x34uy |]
 let private av1C = [| 0x81uy;0x0Cuy;0x3Buy;0x00uy;0x0Auy;0x0Buy;0x77uy;0x88uy |]
-
-/// enhanced 映像 SequenceStart(frameType=1, packetType=0)。
-let private exVideoSeq (fourcc:string) (config:byte[]) =
-    Array.concat [ [| 0x90uy |]; ascii fourcc; config ]
-
-/// enhanced 映像 SequenceStart を ModEx(1バイトの modExData)で包んだもの。
-/// byte0=0x97(frameType=1,packetType=7=ModEx) / 0x00(size-1=0) / 0xAA(modExData) / 0x00(実packetType=0)
-let private exVideoSeqModEx (fourcc:string) (config:byte[]) =
-    Array.concat [ [| 0x97uy; 0x00uy; 0xAAuy; 0x00uy |]; ascii fourcc; config ]
-
-/// enhanced 映像 CodedFrames(packetType=1)。AVC/HEVC は FourCC 直後に符号付き24bit CTS を持つ。
-let private exVideoCodedFrames (fourcc:string) (frameType:int) (cts:int) (payload:byte[]) =
-    let b0 = 0x80 ||| ((frameType &&& 0x07) <<< 4) ||| 0x01
-    Array.concat [ [| byte b0 |]; ascii fourcc; [| byte (cts>>>16); byte (cts>>>8); byte cts |]; payload ]
 
 /// enhanced AV1 CodedFrames(packetType=1)。AV1 は CTS フィールドを持たない(FourCC 直後が即ペイロード)。
 let private exAv1CodedFrames (frameType:int) (obu:byte[]) =
